+++ scripts/copy-dependencies.js
@@ -1,0 +1,9 @@
+import fs from 'fs';
+
+async function main() {
+  fs.copyFileSync('CHANGELOG.md', 'src/assets/CHANGELOG.md');
+}
+
+main().catch((e) => {
+  throw e;
+});