+++ src/app/components/dashboard-navbar/dashboard-navbar.component.ts
@@ -6,6 +6,7 @@ import { GpuAutomationsService } from '../../services/gpu-automations.service';
 import { fade } from '../../utils/animations';
 import { NVMLService } from '../../services/nvml.service';
 import { ElevatedSidecarService } from '../../services/elevated-sidecar.service';
+import { UpdateService } from '../../services/update.service';
 
 @Component({
   selector: 'app-dashboard-navbar',
@@ -16,13 +17,16 @@ import { ElevatedSidecarService } from '../../services/elevated-sidecar.service'
 export class DashboardNavbarComponent implements OnInit {
   settingErrors: Observable<boolean>;
   gpuAutomationsErrors: Observable<boolean>;
+  updateAvailable: Observable<boolean>;
   constructor(
     private settingsService: AppSettingsService,
     private lighthouse: LighthouseService,
     private gpuAutomations: GpuAutomationsService,
     private nvml: NVMLService,
-    private sidecar: ElevatedSidecarService
+    private sidecar: ElevatedSidecarService,
+    private update: UpdateService
   ) {
+    this.updateAvailable = this.update.updateAvailable.pipe(map((a) => !!a.manifest));
     this.settingErrors = combineLatest([
       this.lighthouse.consoleStatus.pipe(
         map((status) => !['UNKNOWN', 'SUCCESS', 'CHECKING'].includes(status)),